refactor: split stage crate into modules
