+++ crates/stages/src/stage.rs
@@ -1,0 +1,97 @@
+use crate::id::StageId;
+use async_trait::async_trait;
+use reth_db::mdbx;
+use reth_primitives::BlockNumber;
+use thiserror::Error;
+
+/// Stage execution input, see [Stage::execute].
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+pub struct ExecInput {
+    /// The stage that was run before the current stage and the block number it reached.
+    pub previous_stage: Option<(StageId, BlockNumber)>,
+    /// The progress of this stage the last time it was executed.
+    pub stage_progress: Option<BlockNumber>,
+}
+
+/// Stage unwind input, see [Stage::unwind].
+#[derive(Debug, PartialEq, Eq, Clone, Copy)]
+pub struct UnwindInput {
+    /// The current highest block of the stage.
+    pub stage_progress: BlockNumber,
+    /// The block to unwind to.
+    pub unwind_to: BlockNumber,
+    /// The bad block that caused the unwind, if any.
+    pub bad_block: Option<BlockNumber>,
+}
+
+/// The output of a stage execution.
+#[derive(Debug, PartialEq, Eq, Clone)]
+pub struct ExecOutput {
+    /// How far the stage got.
+    pub stage_progress: BlockNumber,
+    /// Whether or not the stage is done.
+    pub done: bool,
+    /// Whether or not the stage reached the tip of the chain.
+    pub reached_tip: bool,
+}
+
+/// The output of a stage unwinding.
+#[derive(Debug, PartialEq, Eq, Clone)]
+pub struct UnwindOutput {
+    /// The block at which the stage has unwound to.
+    pub stage_progress: BlockNumber,
+}
+
+/// A stage execution error.
+#[derive(Error, Debug)]
+pub enum StageError {
+    /// The stage encountered a state validation error.
+    ///
+    /// TODO: This depends on the consensus engine and should include the validation failure reason
+    #[error("Stage encountered a validation error in block {block}.")]
+    Validation {
+        /// The block that failed validation.
+        block: BlockNumber,
+    },
+    /// The stage encountered an internal error.
+    #[error(transparent)]
+    Internal(Box<dyn std::error::Error + Send + Sync>),
+}
+
+/// A stage is a segmented part of the syncing process of the node.
+///
+/// Each stage takes care of a well-defined task, such as downloading headers or executing
+/// transactions, and persist their results to a database.
+///
+/// Stages must have a unique [ID][StageId] and implement a way to "roll forwards"
+/// ([Stage::execute]) and a way to "roll back" ([Stage::unwind]).
+///
+/// Stages are executed as part of a pipeline where they are executed serially.
+#[async_trait]
+pub trait Stage<'db, E>: Send + Sync
+where
+    E: mdbx::EnvironmentKind,
+{
+    /// Get the ID of the stage.
+    ///
+    /// Stage IDs must be unique.
+    fn id(&self) -> StageId;
+
+    /// Execute the stage.
+    async fn execute<'tx>(
+        &mut self,
+        tx: &mut mdbx::Transaction<'tx, mdbx::RW, E>,
+        input: ExecInput,
+    ) -> Result<ExecOutput, StageError>
+    where
+        'db: 'tx;
+
+    /// Unwind the stage.
+    async fn unwind<'tx>(
+        &mut self,
+        tx: &mut mdbx::Transaction<'tx, mdbx::RW, E>,
+        input: UnwindInput,
+    ) -> Result<UnwindOutput, Box<dyn std::error::Error + Send + Sync>>
+    where
+        'db: 'tx;
+}
