+++ crates/stages/src/util.rs
@@ -1,0 +1,15 @@
+pub(crate) mod opt {
+    /// Get an [Option] with the maximum value, compared between the passed in value and the inner
+    /// value of the [Option]. If the [Option] is `None`, then an option containing the passed in
+    /// value will be returned.
+    pub(crate) fn max<T: Ord + Copy>(a: Option<T>, b: T) -> Option<T> {
+        a.map_or(Some(b), |v| Some(std::cmp::max(v, b)))
+    }
+
+    /// Get an [Option] with the minimum value, compared between the passed in value and the inner
+    /// value of the [Option]. If the [Option] is `None`, then an option containing the passed in
+    /// value will be returned.
+    pub(crate) fn min<T: Ord + Copy>(a: Option<T>, b: T) -> Option<T> {
+        a.map_or(Some(b), |v| Some(std::cmp::min(v, b)))
+    }
+}
